requirements.txt
pytest==7.4.3
playwright==1.40.0
pytest-playwright==0.4.3
allure-pytest==2.13.2
python-dotenv==1.0.0
flake8==6.1.0
black==23.11.0
pytest-rerunfailures==12.0